pyproject.toml
[project]
name = "habemus-papadum-gcp"
version = "0.1.0-alpha"
description = "Utilities and tools for Google Cloud"
readme = "README.md"
license = { file = "LICENSE" }
keywords = []
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

authors = [
    { name = "Nehal Patel", email = "nehal@alum.mit.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "coolname>=2.2.0",
    "google-api-python-client>=2.186.0",
    "google-auth>=2.42.1",
    "google-cloud>=0.34.0",
    "google-cloud-billing>=1.17.0",
    "google-cloud-iam>=2.15.0",
    "google-cloud-resource-manager>=1.12.0",
    "google-cloud-service-usage>=1.9.0",
    "inquirerpy>=0.3.4",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[project.scripts]
pdum_gcp = "pdum.gcp.cli:main"

[project.urls]
Homepage = "https://github.com/habemus-papadum/gcp"
Repository = "https://github.com/habemus-papadum/gcp"
Documentation = "https://github.com/habemus-papadum/gcp"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdum"]


[dependency-groups]
dev = [
    "hatch>=1.14.2",
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.5.0",
    "playwright>=1.45.0",
    "ruff>=0.13.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.0",
    "mkdocstrings>=0.26.1",
    "mkdocstrings-python>=1.11.1",
    "mkdocs-jupyter>=0.25.1",
    "jupyter>=1.1.1",
    "rich>=13.7.1",
    "inquirerpy>=0.3.4",
    "pre-commit>=3.7.0",
    "watchdog>=3.0.0",
    "marimo>=0.17.3",
]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = []

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-s"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/pdum/gcp"]
relative_files = true
omit = [
    "*/tests/*",
    "*/testing.py",
]
